AIvideo/vibe/requirements.txt
kokoro-onnx
torch
torchvision
torchaudio
diffusers
transformers
accelerate
opencv-python
pillow
numpy
soundfile
